Move OverlayView out of CanvasEditors so Swift can type-check the Mac host.
